+++ poem-openapi/src/lib.rs
@@ -102,6 +102,7 @@
 //! | uuid       | Integrate with the [`uuid` crate](https://crates.io/crates/uuid)|
 //! | url        | Integrate with the [`url` crate](https://crates.io/crates/url) |
 //! | bson        | Integrate with the [`bson` crate](https://crates.io/crates/bson) |
+//! | static-files | Support for static file response |
 
 #![doc(html_favicon_url = "https://raw.githubusercontent.com/poem-web/poem/master/favicon.ico")]
 #![doc(html_logo_url = "https://raw.githubusercontent.com/poem-web/poem/master/logo.png")]
@@ -119,6 +120,7 @@ pub mod param;
 pub mod payload;
 #[doc(hidden)]
 pub mod registry;
+pub mod response;
 pub mod types;
 #[doc(hidden)]
 pub mod validation;
@@ -168,5 +170,5 @@ pub mod __private {
     pub use serde;
     pub use serde_json;
 
-    pub use crate::{base::UrlQuery, payload::ContentTypeTable};
+    pub use crate::base::UrlQuery;
 }