+++ poem-openapi/src/response/mod.rs
@@ -1,0 +1,7 @@
+//! Commonly used response types.
+
+#[cfg(feature = "static-files")]
+mod static_file;
+
+#[cfg(feature = "static-files")]
+pub use static_file::StaticFileResponse;