+++ poem/src/web/mod.rs
@@ -653,7 +653,7 @@ pub struct Html<T>(pub T);
 impl<T: Into<String> + Send> IntoResponse for Html<T> {
     fn into_response(self) -> Response {
         Response::builder()
-            .content_type("text/html")
+            .content_type("text/html; charset=utf8")
             .body(self.0.into())
     }
 }
@@ -838,7 +838,7 @@ mod tests {
         // Html
         let resp = Html("abc").into_response();
         assert_eq!(resp.status(), StatusCode::OK);
-        assert_eq!(resp.content_type(), Some("text/html"));
+        assert_eq!(resp.content_type(), Some("text/html; charset=utf8"));
         assert_eq!(resp.into_body().into_string().await.unwrap(), "abc");
 
         // Json