+++ poem/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.2.49] 2022-1-29
+
+- Make the `StaticFileRequest::create_response` method correctly return `Err(StaticFileError::NotFound)` when the specified file does not exist.
+
+# [1.2.48] 2022-1-27
+
+- The `Content-Type` header of the `Html` response was changed from `text/html` to `text/html; charset=utf8`.
+
 # [1.2.47] 2022-1-26
 
 - Add `TestJsonValue::assert_not_null` method.